+++ gui/admin/src/_components/Footer.svelte
@@ -2,8 +2,8 @@
   <div
     class="container">
     <ul class="grid grid-cols-4 gap-5 text-sm text-center">
-      <li><a class="underline hover:no-underline hover:text-gray-800" href="/gui/graphql">GraphQL Editor</a></li>
-      <li><a class="underline hover:no-underline hover:text-gray-800" href="/gui/liquid">Liquid Evaluator</a></li>
+      <li><a class="underline hover:no-underline hover:text-gray-800" href="http://localhost:3333/gui/graphql">GraphQL Editor</a></li>
+      <li><a class="underline hover:no-underline hover:text-gray-800" href="http://localhost:3333/gui/liquid">Liquid Evaluator</a></li>
       <li><a class="underline hover:no-underline hover:text-gray-800" href="https://documentation.platformos.com">Documentation</a></li>
       <li><a class="underline hover:no-underline hover:text-gray-800" href="https://partners.platformos.com">Partner Portal</a></li>
     </ul>